requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chromedriver-binary==96.0.4664.45.0
cryptography==36.0.0
Django==2.2.5
django-environ==0.8.1
django-glrm==1.1.3
et-xmlfile==1.1.0
h11==0.12.0
idna==3.3
newsapi-python==0.2.6
openpyxl==3.0.9
outcome==1.1.0
psycopg2==2.8.6
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wincertstore==0.2
wsproto==1.0.0
